refactor(mir): split control-flow lowering
Add a 10,000-line ceiling for each lower concern module.
